@@ -1,7 +1,7 @@
 from datetime import datetime
+from typing import TypedDict
 
 from pydantic import BaseModel
-from typing_extensions import TypedDict
 
 
 class BaseAsset(BaseModel):
